REL-291438 - Finally resolved the pipeline build failure! - The new Relativity.Build.Common.targets "AfterBuild" target was conflicted with OI's custom target and wasn't copying the OI binaries correctly. With VS 2017 on forward, the guidance is to no longer use the same target name; rather, use the AfterTargets attribute. - Enhanced the OI test check to make sure the OI directory actually exists before attempting to fetch the list of files. This was the final breakthrough and added better messaging to alert potential future issues.

+++ Source/Relativity.Import.Export.NUnit/OutsideInFileIdServiceTests.cs
@@ -182,23 +182,38 @@ namespace Relativity.Import.Export.NUnit
 				return;
 			}
 
-			string path = OutsideInFileIdService.GetInstallPath();
-			Console.WriteLine($"A serious OI error has occurred. Dumping the list of OI binaries found within the '{path}' install directory.");
-			List<string> files = System.IO.Directory.GetFiles(path).ToList();
-			if (files.Count > 0)
+			try
 			{
-				foreach (string file in files)
+				string path = OutsideInFileIdService.GetInstallPath();
+				Console.WriteLine(
+					$"A serious OI error has occurred. Dumping the list of OI binaries found within the '{path}' install directory.");
+				if (!System.IO.Directory.Exists(path))
 				{
-					Console.WriteLine($"OI binary: {file}");
+					Console.WriteLine($"The OI '{path}' install directory doesn't exist.");
+					Console.WriteLine(
+						"Check the project file and custom OI target to ensure the OI binaries are properly copied to the target path.");
+				}
+				else
+				{
+					List<string> files = System.IO.Directory.GetFiles(path).ToList();
+					if (files.Count > 0)
+					{
+						foreach (string file in files)
+						{
+							Console.WriteLine($"OI binary: {file}");
+						}
+					}
+					else
+					{
+						Console.WriteLine($"No OI binaries were found within the '{path}' install directory.");
+						Console.WriteLine("Check the build scripts and logs for build warnings.");
+					}
 				}
 			}
-			else
+			finally
 			{
-				Console.WriteLine($"No OI binaries were found within the '{path}' install directory.");
-				Console.WriteLine("Check the build scripts logs for build warnings.");
+				dumpBinaries = true;
 			}
-
-			dumpBinaries = true;
 		}
 
 		private void ValidateConfigInfo()
